Added mechanical design files
